+++ src/pages/Cart/Cart.module.scss
@@ -60,158 +60,158 @@ $lightyellow: #fff9ee;
 
 // cardItem
 
-.cartItem {
-  display: flex;
-  width: 100%;
-  border-bottom: 1px solid $bordergray;
+// .cartItem {
+//   display: flex;
+//   width: 100%;
+//   border-bottom: 1px solid $bordergray;
 
-  &:last-child {
-    border-bottom: none;
-  }
-}
+//   &:last-child {
+//     border-bottom: none;
+//   }
+// }
 
-.checkbox {
-  width: 5%;
-  text-align: center;
-  margin-top: 30px;
-  .unchecked {
-    border-radius: 50%;
-    background: #fff;
-    color: #ddd;
-    font-size: 20px;
-    border: 1px solid #ddd;
-    box-sizing: border-box;
-    cursor: pointer;
-  }
-  .checked {
-    border-radius: 50%;
-    background: $yellow;
-    color: #fff;
-    font-size: 20px;
-    box-sizing: border-box;
-    cursor: pointer;
-  }
-}
+// .checkbox {
+//   width: 5%;
+//   text-align: center;
+//   margin-top: 30px;
+//   .unchecked {
+//     border-radius: 50%;
+//     background: #fff;
+//     color: #ddd;
+//     font-size: 20px;
+//     border: 1px solid #ddd;
+//     box-sizing: border-box;
+//     cursor: pointer;
+//   }
+//   .checked {
+//     border-radius: 50%;
+//     background: $yellow;
+//     color: #fff;
+//     font-size: 20px;
+//     box-sizing: border-box;
+//     cursor: pointer;
+//   }
+// }
 
-.infoBox {
-  padding: 30px;
-  border-right: 1px solid $bordergray;
-  position: relative;
-  width: 45%;
-  display: flex;
-  gap: 20px;
-  .productImg {
-    width: 150px;
-    height: 150px;
-    background-color: $gray;
-    box-shadow: 3px 2px 10px $gray;
-    img {
-      width: 100%;
-    }
-  }
-  .productDetail {
-    .name {
-      font-size: 16px;
-      font-weight: 500;
-    }
+// .infoBox {
+//   padding: 30px;
+//   border-right: 1px solid $bordergray;
+//   position: relative;
+//   width: 45%;
+//   display: flex;
+//   gap: 20px;
+//   .productImg {
+//     width: 150px;
+//     height: 150px;
+//     background-color: $gray;
+//     box-shadow: 3px 2px 10px $gray;
+//     img {
+//       width: 100%;
+//     }
+//   }
+//   .productDetail {
+//     .name {
+//       font-size: 16px;
+//       font-weight: 500;
+//     }
 
-    .dueDate {
-      font-size: 14px;
-      font-weight: 100;
-      color: #787878;
-      padding: 10px 0 15px;
-    }
-    .price {
-      font-size: 16px;
-      line-height: 25px;
-    }
+//     .dueDate {
+//       font-size: 14px;
+//       font-weight: 100;
+//       color: #787878;
+//       padding: 10px 0 15px;
+//     }
+//     .price {
+//       font-size: 16px;
+//       line-height: 25px;
+//     }
 
-    .quantityBox {
-      display: flex;
-      align-items: center;
-      padding-top: 15px;
-      gap: 15px;
-      .count {
-        font-size: 13px;
-      }
-      button {
-        width: 24px;
-        height: 24px;
-        line-height: 24px;
-        border-radius: 50%;
-        background-color: $gray;
-        color: #7878;
-        outline: none;
-        border: none;
-        cursor: pointer;
-      }
-    }
-  }
-  .delete {
-    position: absolute;
-    top: 30px;
-    right: 30px;
-    width: 24px;
-    height: 24px;
-    line-height: 24px;
-    color: #787878;
-    text-align: center;
-    font-size: 16px;
-    outline: none;
-    border: 1px solid #f7f7f7f7;
-    background-color: transparent;
-    cursor: pointer;
-  }
-}
+//     .quantityBox {
+//       display: flex;
+//       align-items: center;
+//       padding-top: 15px;
+//       gap: 15px;
+//       .count {
+//         font-size: 13px;
+//       }
+//       button {
+//         width: 24px;
+//         height: 24px;
+//         line-height: 24px;
+//         border-radius: 50%;
+//         background-color: $gray;
+//         color: #7878;
+//         outline: none;
+//         border: none;
+//         cursor: pointer;
+//       }
+//     }
+//   }
+//   .delete {
+//     position: absolute;
+//     top: 30px;
+//     right: 30px;
+//     width: 24px;
+//     height: 24px;
+//     line-height: 24px;
+//     color: #787878;
+//     text-align: center;
+//     font-size: 16px;
+//     outline: none;
+//     border: 1px solid #f7f7f7f7;
+//     background-color: transparent;
+//     cursor: pointer;
+//   }
+// }
 
-.optionBox {
-  width: 35%;
-  display: flex;
-  padding: 30px;
-  align-items: center;
-  border-right: 1px solid $bordergray;
-  ul {
-    width: 100%;
+// .optionBox {
+//   width: 35%;
+//   display: flex;
+//   padding: 30px;
+//   align-items: center;
+//   border-right: 1px solid $bordergray;
+//   ul {
+//     width: 100%;
 
-    li {
-      display: flex;
-      justify-content: space-between;
-      font-size: 15px;
-      color: #222;
-      padding-bottom: 10px;
-      .delete {
-        width: 15px;
-        height: 15px;
-        line-height: 15px;
-        color: #787878;
-        text-align: center;
-        font-size: 10px;
-        outline: none;
-        border: 1px solid #f7f7f7f7;
-        background-color: transparent;
-        cursor: pointer;
-      }
-    }
-  }
-}
+//     li {
+//       display: flex;
+//       justify-content: space-between;
+//       font-size: 15px;
+//       color: #222;
+//       padding-bottom: 10px;
+//       .delete {
+//         width: 15px;
+//         height: 15px;
+//         line-height: 15px;
+//         color: #787878;
+//         text-align: center;
+//         font-size: 10px;
+//         outline: none;
+//         border: 1px solid #f7f7f7f7;
+//         background-color: transparent;
+//         cursor: pointer;
+//       }
+//     }
+//   }
+// }
 
-.priceBox {
-  width: 15%;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  justify-content: center;
-  align-items: center;
-  .price {
-    font-weight: 500;
-  }
-  .delivery {
-    padding: 5px;
-    border-radius: 10px;
-    background-color: $bordergray;
-    font-size: 9px;
-  }
-}
+// .priceBox {
+//   width: 15%;
+//   display: flex;
+//   flex-direction: column;
+//   gap: 10px;
+//   justify-content: center;
+//   align-items: center;
+//   .price {
+//     font-weight: 500;
+//   }
+//   .delivery {
+//     padding: 5px;
+//     border-radius: 10px;
+//     background-color: $bordergray;
+//     font-size: 9px;
+//   }
+// }
 
 .noticeBox {
   padding: 30px;
